refactor(plugin): import ui snapshot types directly
