+++ packages/obsidian-plugin/src/mcp/mcpHttpServiceTypes.ts
@@ -1,0 +1,9 @@
+export type AilssMcpHttpServiceStatusSnapshot = {
+	enabled: boolean;
+	url: string;
+	running: boolean;
+	startedAt: string | null;
+	lastExitCode: number | null;
+	lastStoppedAt: string | null;
+	lastErrorMessage: string | null;
+};
